Solid wireframe material related changes.

+++ contents/shaders/include/common.glsl
@@ -1,5 +1,5 @@
 
-vec2 normalizedToViewport(in vec4 screenRect, in vec2 position)
+vec2 normalizedToViewport(in ivec4 screenRect, in vec2 position)
 {
     return vec2(screenRect.z * 0.5 * (position.x + 1.0) + screenRect.x,
                 screenRect.w * 0.5 * (position.y + 1.0) + screenRect.y);
